d_a_npc_mn: createInit 99.13->99.42 (ret u8->int — retail passes mStts.Init unmasked mr, not clrlwi; remaining = @2100/m_arcname anchor symbols + GetGroundH load order); fresh .o 16:37 postdates

+++ src/d/actor/d_a_npc_mn.cpp
@@ -438,7 +438,7 @@ static s16 daNpcMn_XyCheckCB(void* i_this, int i_param) {
 /* 00000C00-00000F3C       .text createInit__9daNpcMn_cFv */
     /* Nonmatching */
 cPhs_State daNpcMn_c::createInit() {
-    u8 ret = 0xFF;
+    int ret = 0xFF;
     u8 railID = getPrmRailID();
     if (railID != 0xFF) {
         mPathRun.setInf(railID, current.roomNo, 1);
